Custom gate structur option added to the project
Now custom gate structures can be defined for the decompoaition of the unitaries.
Test cases from python and C++ sides added to the project.

+++ decomposition/Decomposition_Base.cpp
@@ -1111,7 +1111,7 @@ int Decomposition_Base::get_operation( unsigned int n, operation_type &type, int
 Operation* Decomposition_Base::get_operation( int n ) {
 
     // get the n-th operation if exists
-    if ( n >= operations.size() ) {
+    if ( (unsigned int) n >= operations.size() ) {
         return NULL;
     }
